CON-1716 Project layout
