+++ helpers/sm_error/not_authorized_error.go
@@ -17,29 +17,10 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package smartling
+package smerror
 
-import (
-	"fmt"
-	"net/url"
-)
+type NotAuthorizedError struct{}
 
-// LimitOffsetRequest is a base request for all other requests to set
-// pagination options,  e.g. limit and offset.
-type LimitOffsetRequest struct {
-	Offset int
-	Limit  int
-}
-
-// GetQuery returns URL-encoded representation of current request.
-func (request LimitOffsetRequest) GetQuery() url.Values {
-	query := url.Values{}
-
-	query.Set("offset", fmt.Sprint(request.Offset))
-
-	if request.Limit > 0 {
-		query.Set("limit", fmt.Sprint(request.Limit))
-	}
-
-	return query
+func (err NotAuthorizedError) Error() string {
+	return "authentication parameters are invalid"
 }
