+++ helpers/sm_file/file_type.go
@@ -17,16 +17,30 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package smartling
+package smfile
 
-// AuthenticationOption specifies should request to API use authentication or
-// not. See Post and Get methods.
-type AuthenticationOption bool
+// FileType represents file type format used in Smartling API.
+type FileType string
 
+// Android and next are types that are supported by Smartling API.
 const (
-	// WithAuthentication equal to use of authentication in request.
-	WithAuthentication = AuthenticationOption(true)
-
-	// WithoutAuthentication equal to not use of authentication in request.
-	WithoutAuthentication = AuthenticationOption(false)
+	FileTypeUnknown        FileType = ""
+	FileTypeAndroid        FileType = "android"
+	FileTypeIOS            FileType = "ios"
+	FileTypeGettext        FileType = "gettext"
+	FileTypeHTML           FileType = "html"
+	FileTypeJavaProperties FileType = "javaProperties"
+	FileTypeYAML           FileType = "yaml"
+	FileTypeXLIFF          FileType = "xliff"
+	FileTypeXML            FileType = "xml"
+	FileTypeJSON           FileType = "json"
+	FileTypeDOCX           FileType = "docx"
+	FileTypePPTX           FileType = "pptx"
+	FileTypeXLSX           FileType = "xlsx"
+	FileTypeIDML           FileType = "idml"
+	FileTypeQt             FileType = "qt"
+	FileTypeResx           FileType = "resx"
+	FileTypePlaintext      FileType = "plaintext"
+	FileTypeCSV            FileType = "csv"
+	FileTypeStringsdict    FileType = "stringsdict"
 )
