+++ helpers/sm_file/file_upload_request.go
@@ -17,7 +17,11 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package smartling
+package smfile
+
+import (
+	"github.com/Smartling/api-sdk-go/helpers/sm_form"
+)
 
 type FileUploadRequest struct {
 	FileURIRequest
@@ -35,7 +39,7 @@ type FileUploadRequest struct {
 	}
 }
 
-func (r *FileUploadRequest) GetForm() (*Form, error) {
+func (r *FileUploadRequest) GetForm() (*sm_form.Form, error) {
 	form, err := r.FileURIRequest.GetForm()
 	if err != nil {
 		return nil, err
