+++ helpers/sm_file/file_upload_result.go
@@ -1,0 +1,7 @@
+package smfile
+
+type FileUploadResult struct {
+	Overwritten bool
+	StringCount int
+	WordCount   int
+}
