+++ helpers/sm_file/limit_offset_request.go
@@ -17,13 +17,29 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package smartling
+package smfile
 
-type FileLastModifiedLocales struct {
-	Items []FileLastModified
+import (
+	"fmt"
+	"net/url"
+)
+
+// LimitOffsetRequest is a base request for all other requests to set
+// pagination options,  e.g. limit and offset.
+type LimitOffsetRequest struct {
+	Offset int
+	Limit  int
 }
 
-type FileLastModified struct {
-	LocaleID     string
-	LastModified UTC
+// GetQuery returns URL-encoded representation of current request.
+func (request LimitOffsetRequest) GetQuery() url.Values {
+	query := url.Values{}
+
+	query.Set("offset", fmt.Sprint(request.Offset))
+
+	if request.Limit > 0 {
+		query.Set("limit", fmt.Sprint(request.Limit))
+	}
+
+	return query
 }
