+++ helpers/utc/utc.go
@@ -17,19 +17,50 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package smartling
+package utc
 
-// File represents a file in the Smartling system.
-type File struct {
-	// FileURI is a unique path to file in Smartling system.
-	FileURI string
+import (
+	"encoding/json"
+	"time"
+)
 
-	// FileType is a file type identifier.
-	FileType FileType
+const utcFormat = "2006-01-02T15:04:05Z"
 
-	// LastUploaded refers to time when file was uploaded.
-	LastUploaded UTC
+// UTC represents time in UTC format (zero timezone).
+type UTC struct {
+	time.Time
+}
 
-	// HasInstructions specifies does files have instructions or not.
-	HasInstructions bool
+// MarshalJSON returns JSON representation of UTC.
+func (utc UTC) MarshalJSON() ([]byte, error) {
+	return json.Marshal(utc.String())
+}
+
+// UnmarshalJSON parses JSON representation of UTC.
+func (utc *UTC) UnmarshalJSON(data []byte) error {
+	var formatted string
+
+	err := json.Unmarshal(data, &formatted)
+	if err != nil {
+		return err
+	}
+
+	location, err := time.LoadLocation("UTC")
+	if err != nil {
+		return err
+	}
+
+	parsed, err := time.ParseInLocation(utcFormat, formatted, location)
+	if err != nil {
+		return err
+	}
+
+	*utc = UTC{parsed}
+
+	return nil
+}
+
+// String returns string representation of UTC.
+func (utc UTC) String() string {
+	return utc.Format(utcFormat)
 }
